Revert `with session(...)` idea back to explicit Header/Terminator sending.
After week of insensitive production usage, idea with hiding processing
Header and Terminator records behind `with` block had been marked as buggy.
Why so?

The main problem comes when this block is used within infinity loop, that pulls
data from LIS or other source. The main use case was looked like:

    def emitter(session):
        for order in query_order_changes():
            if not should_be_processed(order):
                continue
            with session():
                yield make_astm_patient_info(order)
                yield make_astm_order_info(order)

The problem is that after yielding the last record from with block, emitter
processing goes through session.__exit__ setting termination state to let
session been closed. However, since this is non-blocking operation the emitter
workflow goes forward to the loop allowing him emit more and more data till
next yield statement reached. Since there could be no more data to get the whole
process will be blocked and Terminator record will be never send. In this case
it's reasonable behavior for server to drop whole received data since session
transaction commit fails due to timeout error. Such errors are very hard to
handle from user side.

I'd tried many ways to solve this problem, but had found the only worked one
to yield explicitly Header and Terminator records from emitter. This is not
so beautiful solution like it was, but it prevents too much problems to ignore
it. Hope to find the right solution for this later. For now, the
RecordsStateMachine takes care about right message order, so you will be alerted
in case if some record was forgotten to yield.

So, the previous emitter function now should be written as:

    def emitter():
        for order in query_order_changes():
            if not should_be_processed(order):
                continue
            yield Header()
            yield make_astm_patient_info(order)
            yield make_astm_order_info(order)
            yield Terminator()

May be introducing yielding keyword for PEP 3156 is able to solve this problem,
but suddenly it's just possible future solution, not for nowdays.