Apply mod 8 for frame number.
The frame number is an ASCII digit from 0 to 7. Its purpose is to permit
the receiver to distinguish between new and re-transmitted frames. This single
digit is sent immediately after the STX character. The frame number begins with
1 when the transfer phase is initialized and increments by 1 each time a new
frame is transmitted and acknowledged. After 7, the frame number returns to 0
and repeats the above sequence.